Add new file upload constants and move existing constants
