+++ src/main/modules/oidc/index.ts
@@ -4,7 +4,7 @@ import config from 'config';
 import type { Express, NextFunction, Request, Response } from 'express';
 import type * as OidcClientType from 'openid-client';
 
-import { RouteNames } from '../../common-constants';
+import { RouteNames } from '../../constants';
 import type { OIDCConfig } from './config.interface';
 import { OIDCAuthenticationError, OIDCCallbackError } from './errors';
