+++ src/main/routes/generalUpload/index.ts
@@ -9,8 +9,9 @@ import type {
   PreviouslyUploadedDocument,
   PreviouslyUploadedDocumentsCaseData,
 } from '../../app/document/PreviouslyUploadedDocumentClient';
-import { RouteNames } from '../../common-constants';
 import { UploadStepId, uploadSteps } from '../../config/general-upload-config';
+import { RouteNames } from '../../constants';
+import { FileUploadInputFieldNames } from '../../constants/file-upload';
 import { generateRenamedFilename, getCombinedPDFFormat, getDocumentRenameFormat, getSelectedDocumentTypesForDisplay, shouldAutoRename, shouldCombineIntoPDF, toDocumentTypeKey  } from '../../functions/util/documentUtil';
 import { oidcMiddleware } from '../../middleware';
 
@@ -306,6 +307,7 @@ export default function setupGeneralUploadRoute(app: Application): void {
       previousStep,
       caseUserName: req.session.caseUserName,
       contactEmail,
+      FileUploadInputFieldNames: FileUploadInputFieldNames,
       shouldAutoRename,
       getDocumentRenameFormat,
       shouldCombineIntoPDF,
@@ -345,6 +347,7 @@ export default function setupGeneralUploadRoute(app: Application): void {
           values: { selectedDocumentTypes, fdrHearing, uploadMore: req.body.uploadMore, understand: req.body.understand },
           previousStep,
           caseUserName: req.session.caseUserName,
+          FileUploadInputFieldNames: FileUploadInputFieldNames,
           shouldAutoRename,
           getDocumentRenameFormat,
           generateRenamedFilename,
