+++ src/main/routes/info.ts
@@ -3,7 +3,7 @@ import * as os from 'node:os';
 import { infoRequestHandler } from '@hmcts/info-provider';
 import { Router } from 'express';
 
-import { RouteNames } from '../common-constants';
+import { RouteNames } from '../constants';
 
 export default function info(app: Router): void {
   app.get(
