+++ src/test/api/workflows/authentication-flow.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test } from '@jest/globals';
 import request from 'supertest';
 
 import { app } from '../../../main/app';
-import { PrivateRoutes, PublicRoutes } from '../../../main/common-constants';
+import { PrivateRoutes, PublicRoutes } from '../../../main/constants';
 
 describe('Authentication & OIDC Login Flow', () => {
   describe('OIDC Login Flow Initialization', () => {
